Add historical balance behaviour. Add XRP
